+++ src/main/kotlin/org/ratelog/tmdb/TmdbError.kt
@@ -3,6 +3,7 @@ package org.ratelog.tmdb
 import org.ratelog.movie.detail.DetailMovieHandlerError
 import org.ratelog.movie.premieres.MoviePremieresHandlerError
 import org.ratelog.search.SearchHandlerError
+import org.ratelog.search.trending.SearchTrendingHandlerError
 import org.ratelog.tvshow.detail.DetailTvShowHandlerError
 import org.ratelog.tvshow.premieres.TvShowPremieresHandlerError
 
@@ -10,6 +11,7 @@ interface TmdbError:
     DetailMovieHandlerError,
     DetailTvShowHandlerError,
     SearchHandlerError,
+    SearchTrendingHandlerError,
     MoviePremieresHandlerError,
     TvShowPremieresHandlerError {
     object MovieNotFound: TmdbError